Fix: integer type

+++ app/cmd/migrations/sql/000002_create_numeron.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE numerons (
-    `id` bigint(20) unsigned NOT NULL AUTO_INCREMENT COMMENT 'ID',
-    `status` integer unsigned NOT NULL COMMENT 'ステータス',
-    `owner_id` integer unsigned NOT NULL COMMENT '作成者プレイヤーID',
+    `id` integer unsigned NOT NULL AUTO_INCREMENT COMMENT 'ID',
+    `status` tinyint(1) unsigned NOT NULL COMMENT 'ステータス',
+    `owner_id` bigint(20) unsigned NOT NULL COMMENT '作成者プレイヤーID',
     `created_at` TIMESTAMP COMMENT '作成日時',
     `updated_at` TIMESTAMP COMMENT '更新日時',
     `deleted_at` TIMESTAMP COMMENT '更新日時',
